function/schema_image/requirements.txt
google-genai
functions-framework